backend/requirements.txt
flask
# snscrape
# spacy
# spacy-langdetect
# spacy-streamlit>=1.0.0a0
# streamlit
pymongo[srv]
python-dotenv